Assignment@/image_processing_backend/requirements.txt
Django==4.2
djangorestframework==3.14
celery==5.2.7
redis==4.5.1
#psycopg2-binary==2.9.6
Pillow==9.5.0
drf-yasg==1.20.0  # Swagger API documentation
python-dotenv==1.0.0  # For environment variables
